Projects/T13_Color_Transfer_using_GUI/requirements.txt
streamlit
Pillow
opencv-python
numpy
matplotlib